# Backend/src/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SchooliAT API",
    "description": "School Management Application API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ExamType": {
        "type": "string",
        "enum": ["UNIT_TEST", "SEMESTER", "FINAL"]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/exams": {
      "post": {
        "tags": ["Exams"],
        "summary": "Create a new exam",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["request"],
                "properties": {
                  "request": {
                    "type": "object",
                    "required": ["year", "name", "type"],
                    "properties": {
                      "year": {
                        "type": "integer",
                        "minimum": 1900,
                        "maximum": 2100
                      },
                      "name": { "type": "string" },
                      "type": { "$ref": "#/components/schemas/ExamType" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Exam created" }
        }
      },
      "get": {
        "tags": ["Exams"],
        "summary": "Get all exams",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Exams fetched" }
        }
      }
    },
    "/exams/{id}": {
      "patch": {
        "tags": ["Exams"],
        "summary": "Update an exam",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "request": {
                    "type": "object",
                    "properties": {
                      "year": {
                        "type": "integer",
                        "minimum": 1900,
                        "maximum": 2100
                      },
                      "name": { "type": "string" },
                      "type": { "$ref": "#/components/schemas/ExamType" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Exam updated" },
          "404": { "description": "Exam not found" }
        }
      },
      "delete": {
        "tags": ["Exams"],
        "summary": "Delete an exam",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Exam deleted" },
          "404": { "description": "Exam not found" }
        }
      }
    },
...
    "/files": {
      "post": {
        "tags": ["Files"],
        "summary": "Upload a file",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "File uploaded" },
          "400": { "description": "File is required" }
        }
      }
    },
    "/files/{id}": {
      "get": {
        "tags": ["Files"],
        "summary": "Get a file by ID (development only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "File content" },
          "404": { "description": "File not found" }
        }
      }
    },
...
    "/locations": {
      "post": {
        "tags": ["Locations"],
        "summary": "Create a location",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["request"],
                "properties": {
                  "request": {
                    "type": "object",
                    "required": ["name", "regionId", "employeeId"],
                    "properties": {
                      "name": { "type": "string" },
                      "regionId": { "type": "string", "format": "uuid" },
                      "employeeId": { "type": "string", "format": "uuid" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Location created" }
        }
      },
      "get": {
        "tags": ["Locations"],
        "summary": "Get all locations",
        "parameters": [
          {
            "name": "employeeId",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "regionId",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "schema": { "type": "integer" }
          },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Locations fetched" }
        }
      }
    },
    "/locations/{id}": {
      "delete": {
        "tags": ["Locations"],
        "summary": "Delete a location",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Location deleted" }
        }
      }
    },
...
  },
...
}